+++ script/workspace/require-path.lua
@@ -111,14 +111,6 @@ function m.findUrisByRequirePath(suri, path)
     tracy.ZoneBeginN('findUrisByRequirePath')
     local results = {}
     local searchers = {}
-    for uri in files.eachDll() do
-        local opens = files.getDllOpens(uri) or {}
-        for _, open in ipairs(opens) do
-            if open == fspath then
-                results[#results+1] = uri
-            end
-        end
-    end
 
     ---@type collector
     local clt = scope.getScope(suri):get('requireName')
@@ -133,6 +125,15 @@ function m.findUrisByRequirePath(suri, path)
                         searchers[uri] = info.searcher
                     end
                 end
+            end
+        end
+    end
+
+    for uri in files.eachDll() do
+        local opens = files.getDllOpens(uri) or {}
+        for _, open in ipairs(opens) do
+            if open == fspath then
+                results[#results+1] = uri
             end
         end
     end