+++ changelog.md
@@ -5,6 +5,7 @@
   * `type.weakNilCheck`
 * `FIX` [#1256](https://github.com/sumneko/lua-language-server/issues/1256)
 * `FIX` [#1257](https://github.com/sumneko/lua-language-server/issues/1257)
+* `FIX` [#1267](https://github.com/sumneko/lua-language-server/issues/1267)
 * `FIX` [#1269](https://github.com/sumneko/lua-language-server/issues/1269)
 * `FIX` [#1273](https://github.com/sumneko/lua-language-server/issues/1273)
 * `FIX` [#1275](https://github.com/sumneko/lua-language-server/issues/1275)